1. Fixed tools_def.template to meet ICC build for IA32 2. Modified some source files to meet ICC build for IA32 and IPF.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@3271 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Bus/Pci/EhciDxe/EhciUrb.c
@@ -486,7 +486,7 @@ EhcCreateQtds (
     // Switch the Toggle bit if odd number of packets are included in the QTD.
     //
     if (((Qtd->DataLen + Ep->MaxPacket - 1) / Ep->MaxPacket) % 2) {
-      Toggle = 1 - Toggle;
+      Toggle = (UINT8) (1 - Toggle);
     }
 
     Len += Qtd->DataLen;
@@ -588,7 +588,7 @@ EhcCreateUrb (
 
   Ep              = &Urb->Ep;
   Ep->DevAddr     = DevAddr;
-  Ep->EpAddr      = EpAddr & 0x0F;
+  Ep->EpAddr      = (UINT8) (EpAddr & 0x0F);
   Ep->Direction   = ((EpAddr & 0x80) ? EfiUsbDataIn : EfiUsbDataOut);
   Ep->DevSpeed    = DevSpeed;
   Ep->MaxPacket   = MaxPacket;
